move input module items to the lower module
