+++ crates/ra_hir/src/query_definitions.rs
@@ -6,16 +6,14 @@ use std::{
 use rustc_hash::FxHashMap;
 use ra_syntax::{
     AstNode, SyntaxNode, TreeArc,
-    ast::{self, ModuleItemOwner}
 };
 use ra_db::SourceRootId;
 
 use crate::{
-    SourceFileItems, SourceItemId, DefId, HirFileId, ModuleSource,
-    MacroCallLoc, FnScopes,
+    SourceFileItems, SourceItemId, DefId, HirFileId,
+    FnScopes,
     db::HirDatabase,
-    module_tree::ModuleId,
-    nameres::{InputModuleItems, ItemMap, Resolver},
+    nameres::{ItemMap, Resolver},
 };
 
 pub(super) fn fn_scopes(db: &impl HirDatabase, def_id: DefId) -> Arc<FnScopes> {
@@ -41,56 +39,6 @@ pub(super) fn file_item(
             .syntax()
             .to_owned(),
     }
-}
-
-pub(super) fn input_module_items(
-    db: &impl HirDatabase,
-    source_root_id: SourceRootId,
-    module_id: ModuleId,
-) -> Arc<InputModuleItems> {
-    let module_tree = db.module_tree(source_root_id);
-    let source = module_id.source(&module_tree);
-    let file_id = source.file_id;
-    let source = ModuleSource::from_source_item_id(db, source);
-    let file_items = db.file_items(file_id);
-    let fill = |acc: &mut InputModuleItems, items: &mut Iterator<Item = ast::ItemOrMacro>| {
-        for item in items {
-            match item {
-                ast::ItemOrMacro::Item(it) => {
-                    acc.add_item(file_id, &file_items, it);
-                }
-                ast::ItemOrMacro::Macro(macro_call) => {
-                    let item_id = file_items.id_of_unchecked(macro_call.syntax());
-                    let loc = MacroCallLoc {
-                        source_root_id,
-                        module_id,
-                        source_item_id: SourceItemId {
-                            file_id,
-                            item_id: Some(item_id),
-                        },
-                    };
-                    let id = loc.id(db);
-                    let file_id = HirFileId::from(id);
-                    let file_items = db.file_items(file_id);
-                    //FIXME: expand recursively
-                    for item in db.hir_source_file(file_id).items() {
-                        acc.add_item(file_id, &file_items, item);
-                    }
-                }
-            }
-        }
-    };
-
-    let mut res = InputModuleItems::default();
-    match source {
-        ModuleSource::SourceFile(it) => fill(&mut res, &mut it.items_with_macros()),
-        ModuleSource::Module(it) => {
-            if let Some(item_list) = it.item_list() {
-                fill(&mut res, &mut item_list.items_with_macros())
-            }
-        }
-    };
-    Arc::new(res)
 }
 
 pub(super) fn item_map(db: &impl HirDatabase, source_root: SourceRootId) -> Arc<ItemMap> {
